requirements.txt
Flask
Flask_sqlalchemy
python-dotenv
requests
gunicorn
